Copied old Match class to StrongMatch. Removed Bulk of StrongMatch code (everything that could be considered 'specific' to FIRST Stronghold)
